DRIVER NOTIFICATION BOT/dashboard/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
asyncpg>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
